refine command / response protocol
It turns out that protobuf does not include the type of a message in the
message, so you can't just parse through a bunch of messages of mixed
types. Instead, the documentation[1] suggests a container message that
includes each of the valid messages as optionals, and then expecting
only that message type.

Introduce a new message type that contains all the valid commands that
could be sent, and simplify the responses into a single message.

[1] https://developers.google.com/protocol-buffers/docs/techniques#union