Add to documentation and Makefile
The readme didn't describe some of the dependencies that are necessary
to compile the C part, and run the Python part. Also, the steps for
getting the protobuf file generated for Python was not documented or
automated. Also, there was no simple way to clean up the generated
files. All of this is now fixed.